I'm not hurt Yonero
